refactor(push): separate registrations from delivery
